+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (5.0.0) IN PROGRESS
 * [PERF-62](https://issues.folio.org/browse/PERF-62) Use more efficient queries.
 * Refactor to `miragejs` from `bigtest/mirage`.
+* [UITEN-105](https://issues.folio.org/browse/UITEN-105) Settings > Tenant > Locale show locales in current and native locale
 * Increment `@folio/stripes` to `v5`, `react-router` to `v5.2`.
 
 ## [4.0.0](https://github.com/folio-org/ui-organization/tree/v4.0.0) (2020-06-11)